frontend/requirements.txt
fastapi>=0.111.0
python-dotenv>=1.0.0
redis[hiredis]>=5.0.0
sse-starlette>=1.8.0
uvicorn[standard]>=0.29.0
